move tx-config to core
